fix(money): widen Int32 to Int64 on price/value fields
Sweep of stragglers missed in iter-2: PointValue on Sale/SaleItem,
PriceOverride on ProductBarcode and Discount. Repositories updated to
read via GetLong so values above Rp 21.4M do not silently overflow.

+++ Kasir.Core/Data/Repositories/DiscountRepository.cs
@@ -52,7 +52,7 @@ namespace Kasir.Data.Repositories
                 DateEnd = SqlHelper.GetString(reader, "date_end"),
                 MinQty = SqlHelper.GetInt(reader, "min_qty"),
                 MaxQty = SqlHelper.GetInt(reader, "max_qty"),
-                PriceOverride = SqlHelper.GetInt(reader, "price_override"),
+                PriceOverride = SqlHelper.GetLong(reader, "price_override"),
                 Description = SqlHelper.GetString(reader, "description"),
                 Priority = SqlHelper.GetInt(reader, "priority"),
                 IsActive = SqlHelper.GetInt(reader, "is_active"),
